+++ src/fields/formfields/BaseOptionsField.php
@@ -11,6 +11,7 @@ use craft\fields\BaseOptionsField as CraftBaseOptionsField;
 use craft\fields\data\MultiOptionsFieldData;
 use craft\fields\data\OptionData;
 use craft\fields\data\SingleOptionFieldData;
+use craft\helpers\Json;
 
 abstract class BaseOptionsField extends CraftBaseOptionsField
 {