+++ src/fields/formfields/Entries.php
@@ -46,6 +46,7 @@ class Entries extends CraftEntries implements FormFieldInterface
         getSavedFieldConfig as traitGetSavedFieldConfig;
         getSettingGqlTypes as traitGetSettingGqlTypes;
         RelationFieldTrait::defineValueAsString insteadof FormFieldTrait;
+        RelationFieldTrait::defineValueAsJson insteadof FormFieldTrait;
         RelationFieldTrait::getIsFieldset insteadof FormFieldTrait;
         RelationFieldTrait::populateValue insteadof FormFieldTrait;
         RelationFieldTrait::renderLabel insteadof FormFieldTrait;