+++ src/fields/formfields/Password.php
@@ -186,4 +186,12 @@ class Password extends FormField implements PreviewableFieldInterface
 
         return '';
     }
+
+    /**
+     * @inheritDoc
+     */
+    protected function defineValueAsJson($value, ElementInterface $element = null)
+    {
+        return $this->defineValueAsString($value, $element);
+    }
 }