+++ src/fields/formfields/Tags.php
@@ -46,6 +46,7 @@ class Tags extends CraftTags implements FormFieldInterface
         getSettingGqlTypes as traitGetSettingGqlTypes;
         getDisplayTypeValue as traitGetDisplayTypeValue;
         RelationFieldTrait::defineValueAsString insteadof FormFieldTrait;
+        RelationFieldTrait::defineValueAsJson insteadof FormFieldTrait;
         RelationFieldTrait::getIsFieldset insteadof FormFieldTrait;
         RelationFieldTrait::populateValue insteadof FormFieldTrait;
         RelationFieldTrait::renderLabel insteadof FormFieldTrait;